requirements.txt
amqp==5.2.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
Brotli==1.1.0
cairocffi==1.6.1
CairoSVG==2.7.1
celery==5.2.7
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
csscompressor==0.9.5
cssselect2==0.7.0
defusedxml==0.7.1
dj-rest-auth==2.1.12
Django==4.1
django-allauth==0.51.0
django-compress-staticfiles==1.0.1b0
django-extensions==3.1.5
django-filter==21.1
django-ipware==3.0.7
django-model-utils==4.2.0
django-organizations==2.0.2
django-phonenumber-field==6.0.0
django-private-storage==3.1.1
django-sendsms==0.5
django-sesame==2.4
djangorestframework==3.13.1
djangorestframework-link-header-pagination==0.1.1
drf-yasg==1.20.3
html5lib==1.1
idna==3.7
importlib-metadata==4.13.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsonfield==3.1.0
kombu==5.3.7
MarkupSafe==2.1.5
oauthlib==3.2.2
openwisp-radius==1.0.2
openwisp-users==1.0.2
openwisp-utils==1.0.5
packaging==24.0
passlib==1.7.4
phonenumbers==8.12.57
pillow==10.3.0
prompt-toolkit==3.0.43
pycparser==2.22
PyJWT==2.8.0
pyphen==0.14.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rjsmin==1.2.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
sqlparse==0.5.0
swapper==1.3.0
tinycss2==1.2.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
WeasyPrint==52.5
webencodings==0.5.1
zipp==3.18.1
